# api_gateway/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/create/student_task": {
            "post": {
                "description": "API for creating a student task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student_task"
                ],
                "summary": "Create student task",
                "parameters": [
                    {
                        "description": "Student Task",
                        "name": "student_task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/education_management_service.CreateStudentTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/education_management_service.StudentTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/student_task/list": {
            "get": {
                "description": "API for getting a list of student tasks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student_task"
                ],
                "summary": "Get list of student tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/education_management_service.GetListStudentTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/student_task/score/student/{id}": {
            "put": {
                "description": "API for updating a student task score for a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student_task"
                ],
                "summary": "Update student task score for student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Student Task Score",
                        "name": "score",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/education_management_service.UpdateStudentScoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/education_management_service.GetStudentTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Student Task not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/student_task/score/teacher/{id}": {
            "put": {
                "description": "API for updating a student task score for a teacher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student_task"
                ],
                "summary": "Update student task score for teacher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Student Task Score",
                        "name": "score",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/education_management_service.UpdateStudentScoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/education_management_service.GetStudentTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Student Task not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/student_task/student/{id}": {
            "get": {
                "description": "API for getting a student task by student ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student_task"
                ],
                "summary": "Get student task by student ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/education_management_service.GetStudentTaskResponse"
                        }
                    },
                    "404": {
                        "description": "Student Task not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/student_task/{id}": {
            "get": {
                "description": "API for getting a student task by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student_task"
                ],
                "summary": "Get student task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/education_management_service.GetStudentTaskResponse"
                        }
                    },
                    "404": {
                        "description": "Student Task not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "API for updating a student task by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student_task"
                ],
                "summary": "Update student task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Student Task",
                        "name": "student_task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/education_management_service.UpdateStudentTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/education_management_service.GetStudentTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Student Task not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "API for deleting a student task by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student_task"
                ],
                "summary": "Delete student task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/education_management_service.StudentTaskEmpty"
                        }
                    },
                    "404": {
                        "description": "Student Task not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "education_management_service.CreateStudentTaskRequest": {
            "type": "object",
            "properties": {
                "score": {
                    "type": "integer"
                },
                "studentId": {
                    "type": "string"
                },
                "taskId": {
                    "type": "string"
                }
            }
        },
...
        "education_management_service.GetListStudentTaskResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "student_tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/education_management_service.GetStudentTaskResponse"
                    }
                }
            }
        },
...
        "education_management_service.GetStudentTaskResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "studentId": {
                    "type": "string"
                },
                "taskId": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "education_management_service.StudentTaskEmpty": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "education_management_service.StudentTaskResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "studentId": {
                    "type": "string"
                },
                "taskId": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "education_management_service.UpdateStudentScoreRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "education_management_service.UpdateStudentTaskRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "studentId": {
                    "type": "string"
                },
                "taskId": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseError": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}